split the ES and Opensearch Utility into 2 classes
fixed bug in _pit where size defaults to 10 when not given
added opensearch-py to setup.py
bumped minor version to 1.1.0
