+++ setup.py
@@ -23,6 +23,7 @@ setup(
     zip_safe=False,
     install_requires=[
         'elasticsearch>=7.0.0,<7.14.0',
+        'opensearch-py>=2.3.0,<3.0.0',
         'requests>=2.7.0',
         'future>=0.17.1',
         "jsonschema>=3.0.1",
